Remove donation status from presenters
We don't use the donation status in the Fundraising App frontend code.
Removed some status constants used in Version 6 of the Donation bounded
context (a BC break).
Instead of removing the mapped constants, we remove the output of the
status completely.

Changed one test to avoid calling Donation::getStatus().

This is the final piece for preparing
https://phabricator.wikimedia.org/T281853

When this commit is merged, we can safely remove `status`-related code
from the `Domain\Model\Donation` in the donation bounded context.