Suppress deprecations warnings for MCP
Version <=1.25 of the Mcp library uses default parameters before
parameters without defaults, which is deprecated in PHP 8. This causes
20+ deprecation warnings during the unit tests. To keep the unit test
output useful, we suppress the warnings for now (and have notified the
provider of the library).

When the deprecation has been fixed, revert this commit.

This is for https://phabricator.wikimedia.org/T270715