Refactor Twig factories
Split filters and functions into getters in the abstract class and
return them from the child classes. This is a better structure because
child classes can rely on the base class when they don't have filters or
functions.

Pass in dependencies in constructor

Remove locale parameter from constructor, because the locale is not used
anywhere in the factories. Instead, the locales will affect the
dependencies that are passed in.

This is for https://phabricator.wikimedia.org/T284908