Replace FilePrefixer with Symfony Asset
Use the `asset` function in templates to define paths to assets (both to
skin-related assets and assets that come with the application).

The new class AssetPackageFactory defines the packages of assets. In
this commit they are static, but future versions of the factory will use
the "asset-path" setting from the configuration to re-enable loading
assets from a development server URL.

In future versions AssetPackageFactory will also add cache-busting
prefixes to the assets, like FilePrefixer did.

This is for https://phabricator.wikimedia.org/T284908