Use manifest file for assets
The manifest will translate file paths to file paths with cache-busting
hashes. We're not using it in the test environment because the test
environment should be independent from actual asset files.

This commit makes the pull request dependent on
https://gitlab.com/fun-tech/fundraising-app-frontend/-/merge_requests/89
which introduces the manifest file.

This is for https://phabricator.wikimedia.org/T284908