Remove TLS flag from Mailer instantiation
This will make the mailer "auto-discover" the encryption capabilities of
the SMTP server, allowing us to use port 587 with STARTTLS negotiation.
Another solution would be to use port 465 with TLS set to `true`,
but the [recommended way][1] is to use port 587 and let the library
negotiate the encryption.

See also https://github.com/symfony/symfony/issues/34846#issuecomment-580190302

[1]: https://sendgrid.com/blog/whats-the-difference-between-ports-465-and-587/