Add Improve ContextFactoryCollection
Remove IteratorAggregate interface - we're no longer iterating over
context factories. If we need to iterate over a subset (e.g. calling
`registerCustomTypes` on factories that support it, we'll add another
method to ContextFactoryCollection)

Ignore code coverage for getting XML from on legacy factories - the
code is tested implicitly through our route tests and testing it
explicitly would lead to breakages whenever we test a factory where we
added the ``getDoctrineMappingPaths` method

Modernize `BucketTestingContextFactory`

Trigger deprecation notice when all factories have been ported.
I was not successful in expecting the notice in the test through
$this->expectDeprecation(). The deprecation notice is shown, but PHPUnit
doesn't catch it with its error handler. I opted for suppressing the
notice in the non-canary test instead.