Allow empty payment type for fee validation
Currently, the frontend does not send a payment type (because until we
implement https://phabricator.wikimedia.org/T312070 we have only one
payment type), but the new payment domain requires a payment type when
validating (because it's trying to construct a valid Payment
internally).

This problem only occurs in the fee validation. For applying for
memberships the frontend sends a type. For donations we don't have
server-side validation of the amount.