Revert "Merge pull request #2602 from wmde/quick-fix-ppl-encoding"
The encoding of the PayPal data from ISO-8859-1 to UTF-8 broke the
verification service that sends the transaction data back to Paypal to
see if it originated with PayPal and was not tampered with.

We'll solve this further commits by using a modified Payment bounded
content that ignores more transaction fields when writing to the
database.

This reverts commit 1cd3f8c0d50e59055ce352dce742c2ab3b200d41, reversing
changes made to 81e965e0c76fae7e19922dc33a5e031c579de3ee.