Inject native Symfony mailer in FunFunFactory
Add setter for Symfony mailer in FunFunFactory and remove creating a
SMTP mailer transport from the configuration.

Add default mailer configuration to the framework configuration. We're
using the "Native mailer" (a wrapper for the PHP `mail` command). In
production we have a postfix setup that will forward the emails to
email.wikipedia.de. In our dev environment, the sendmail replacement in
our Docker containers will deliver the mail to the MailHog container.

Ticket: https://phabricator.wikimedia.org/T321559