Avoid PayPal error-retry loops for booked payments
Check booking response if the failure reason is that the payment was
already booked. We still log the error (as a warning), but no longer
return an error response, because PayPal would try to book the
already-booked payment again several times.

Check success response for empty body in the tests.

Ticket:  https://phabricator.wikimedia.org/T321346