Remove micropayment service client
The vendor has pulled the 1.25 binary, which led to errors on
installation.

We no longer use the library, so we can safely remove it.

Previously, we used it to fetch the expiry date of a credit card when we
got a credit card notification, but we removed this functionality for
PCI compliance and data protection reasons when we upgraded the payment
bounded context (Ticket https://phabricator.wikimedia.org/T305827)

Removing a library made it necessary to update all dependencies to
regenerate composer.lock.