Fix unit tests for PHPUnit 10
Remove deprecated `assertObjectHasAttribute` calls.
See https://github.com/sebastianbergmann/phpunit/issues/4601 for the
deprecation decision.

In some cases, we can rely on PHP errors when the attribute is not
present, in other cases (where errors would be very confusing or where
the existence of the property is the only test) we implement our own
check and message.