Remove deprecation notice in tests
In previous versions of PHPUnit we were able to suppress a canary
deprecation notice in our code. Now we need to avoid generating that
deprecation notice because it's a false positive.

The actual canary test that checks for the full migration of the
deprecated feature stays as it is. If it starts throwing exceptions, we
can refactor ContextFactoryCollection.