Revert "Merge pull request #2764 from wmde/informal-active"
We need to deploy the changes for "remove WMFG" (special branch in the
content repo), but the informal membership email templates are not
ready.

This revert is temporary, the branch this commit is in should not be
merged!

This reverts commit 1d31239526c96ac28940accba0e649a19e4d826b, reversing
changes made to 298772398d9ca26d6f1fb7206e45f0b762de6e45.