Allow null intervals in initialFormValues
Interval was previously always being converted
with intval before being output to the front
end. Because it always had a value it caused
the interval in local storage on the form to
be overwritten on page refresh.

This change allows it to be null.

Ticket: https://phabricator.wikimedia.org/T342547