Cache the PayPal API configuration
The PayPal API configuration is in a YAML file that gets read and
validated by `PayPalURLGeneratorConfigReader`. To avoid parsing and
validating on every request, we cache the output of the class in
`PayPalAPIURLGeneratorConfigLoader`.

We realized that we can reuse the cache for the campaign configuration,
because the "source" files for the configuration are in the same path
(`app/config`), so the cache keys won't collide. Renamed "CampaignCache"
to "ConfigurationCache".

Co-Authored-By: Abban Dunne <abban.dunne@wikimedia.de>
Co-Authored-By: Corinna Hillebrand <corinna.hillebrand@wikimedia.de>
Co-Authored-By: Sperling-0 <tanuja.doriya@wikimedia.de>