Set http_method_override in config
In Symfony 7.0 http_method_override will be a required value. This
change removes deprecation warnings we are getting in environments
where APP_DEBUG=0

Setting it to `false` will forbid the trick of passing a `_method`
parameter to in the URL to override the HTTP method. We don't use this
in the frontend code and can safely disable it.