Add Authentication classes
This commit implements the authentication interfaces for the bounded
contexts, with persistence. I called the implementation `OldStyleTokens`
to leave the option to have a different implementation that uses only
tokens.

This also removes the reliance on the Doctrine event system for
generating tokens. All the code in the Context Factories in the bounded
contexts can be removed.

I accidentally force-pushed to the bounded context branches, which led
to errors during the rebase of the `payment-with-ppl` branch. I removed
the composer.lock during the rebase and I'm adding it back here. This
commit fixes the broken `payment-with-ppl` branch.

Tickets:
https://phabricator.wikimedia.org/T341795
https://phabricator.wikimedia.org/T344271