Append bucket testing parameters to redirect URLs
Add an HTTP response event handler that appends the bucket testing
parameters to all redirect URLs that lead to other routes in the
Fundraising Application. This preserves the bucket selection between
pages, without resorting to a PHP session or client-side storage, both
of which would require user consent.

This change is important for banner test C23_WMDE_Desktop_DE_06 (tracked
in ticket [T345721](https://phabricator.wikimedia.org/T345721)), but is
useful for all subsequent tests.

This change does NOT add the parameters to the confirmation page for
external payments - we'd have to expose the URL generation for payment
providers to the app via interface/events. With the changes coming in
the PayPal API implementation tickets (see [parent ticket
T329159](https://phabricator.wikimedia.org/T329159) ) it would be
duplicated and increased effort to do it now.

Ticket: https://phabricator.wikimedia.org/T347008