Validate default development config
To avoid a broken out-of-the-box configuration, validate the default
development configuration in `.docker/app/config.dev.json` (copied to
`app/config` by the `make default-config` task). This will remind us to
change both the default configuration and the test configuration
whenever we make changes to the configuration schema in
`app/config/schema.json`.

Remove the siteUrlBase property from the required Matomo configuration
settings - it was used for the server-side tracking, which was removed
in Nov 2020 with commit f28cfd90f97ec4d971f247e8af42d3c7d5642309

Rename "Piwik" to "Matomo", at least in the item descriptions.

This is a change that'll affect deployment, please adapt the
configuration files in the infrastructure repository