Improve PayPal config handling
Check if PayPal API config file exists before trying to read it. This
does not help much if the `config-path` setting is empty, but that
should be caught by validating the configuration file.

Remove call to `cache->has` because the docs recommend using it only for
cache warming, not for checking.