Rebuild Drone cache before running tests
Our CI fails randomly, with an error that files are missing.
This looks like it's coming from a phpStan cleanup that's
happening simultaneously with the cache rebuild step.

This commit moves the rebuild cache step between the build and
test steps so it will rebuild the cache before Stan runs.