Change cache location in dev and test environments
Do not rely on a mounted volume, use the `/tmp` path instead.
We choose `/tmp` because it's world-writable and the `www-data` process
(PHP-FPM) needs to write there.

This simplifies our Docker image where we don't have to create the mount
point `/usr/share/nginx/www/spenden.wikimedia.de/current` with the right
permissions. We also don't need an extra volume any more. Removing the
Docker container will also remove the data in the cache.

Improve README formatting and language and add instructions for clearing
the cache.