Fix overriding application config in tests
While running `ShowUpdateAddressFormRouteTest` individually, I ran into
the bug that one test was failing (because a template file was not
present in `skins/test` but was is present in `skins/laika`), but only
when I did run it individually, not when running `make test`. The
cause of this was as follows:

`WebRouteTestCase` calls `bootKernel` which initializes all Symfony
dependencies. One of those dependencies is the `SkinCacheWarmer`, which
has `FunFunFactory` as a dependency, creating a Twig instance. This
dependency causes Symfony to initialize `FunFunFactory` with the default
configuration for the test environment, using `test` as the skin name.
The mechanism of `WebRouteTestCase` for overriding the configuration
passed to `FunFunFactory` assumes that the `bootKernel` method won't
initialize `FunFunFactory`.

The "temporary" solution is to move the cache warmer out of the testing
dependencies and into the production dependencies.

A better solution would be to initialize it without the `FunFunFactory`,
but the Twig environment has a lot of other dependencies that we'd have
to move out of the `FunFunFactory` and into `services.yml` that we'll
probably live with the current solution for some time.

Another solution would be change how we override the skin name.

When we've either changed the application to provide API for a single
page application (Ticket https://phabricator.wikimedia.org/T241751 ) or
removed the dependency of our tests on the `test` skin
(Ticket https://phabricator.wikimedia.org/T241751 ) we can move the
cache warming service back into `services.yml` because the default
skin in the configuration won't need overrides in the tests.