Add API controller for donation comments
The new frontend code now sends a JSON body, so we can take the
opportunity to modernize the API.

Use `#[MapRequestPayload]` attribute for JSON serialization and type
checking.

Reused most of the controller and test code of the old controller, which
can be deleted when the frontend is merged.

Updated all dependencies when adding new dependency
`symfony/serializer-pack` .

Ticket: https://phabricator.wikimedia.org/T367387