Improve PayPal notifications
We have a notification where the `item_number` (that can contain the
donation ID) is present, but an empty string. We can't use
`ParameterBag::getInt` in this case, because it strictly validates the
number and an empty string will cause the validation to fail.
Instead, we now use `ParameterBag::filter`, which handles both
cases (empty and invalid).

See https://www.php.net/manual/en/function.filter-var.php for more
information (`ParameterBag::filter` is a wrapper around `filter_var`
with additional checks).