Update to PHPUnit 11
We were held back by https://github.com/symfony/symfony/issues/53812
The issue is that `FrameworkBundle` registers an exception handler (if
`ErrorHandler` was not registered before as an exception handler) and
never removes it when the Kernel shuts down. PHPUnit checks for newly
registered exception handlers at the end of each test, so whenever the
kernel boots, PHPUnit sees a "dangling" exception handlers and issues a
warning.

Now that the issue is closed with a documented workaround (but not a
proper solution in `FrameworkBundle`) the only way forward seems to be
the workaround. "Workaround" means registering the Symfony error handler
*before* starting PHPUnit, which will skip the additional registration
in `FrameworkBundle`.

Ticket: https://phabricator.wikimedia.org/T359971