Improve output of JSON validation errors
When using the `#[MapRequestPayload]` attribute, Symfony will throw a
`ValidationFailedException` if the JSON sent by the client does not
match the expected class definition (e.g. missing field, fields of wrong
type, etc). This commit adds an `validationErrors` field to the
resulting JSON with `field name => error` pairs to help the developer
debug the fields better.

We keep the `ERR` property in the result as an "indicator" value for
backwards compatibility.