Fix call to json_decode in ConfigReader
The call was missing a parameter, which could have led to a non-object
return value. But since we pass in valid JSON into `json_decode` and
`JSON_THROW_ON_ERROR` has a high integer value that could be applied as
the `depth` value for `json_decode`, the missing parameter never became
an issue.